requirements.txt
pyusb==1.3.1
libusb-package==1.0.26.3
hidapi==0.15.0
